Added more code.
I need input on the LED placement and the purpose of each LED strip. Need some ideas to what we can use the LEDs for.. So far, we have two LED strips (8 LEDs on each) wired in parallel on Intake arm for "feeding" to let the driver and other teammates know when we are able to grab another ball after launching it. Any other input would be appreciated. Thanks.