Modified ball detector stuff.
Added a new threshold value to RobotMap. The idea is to have the sensor have one value which determines if a ball is present within the robot. This will trigger the ball tappers to a position to help settle the ball. Another threshold will determine that the ball has settled on the catapult.

Added some commands that wait until a ball is detected, and wait until a ball is not detected. These two states are required to step through a sequence of ball acquisition or deployment.

Need to think about this some more, but we might need two seperate sensors. THe code isn't currently set up to use two sensors. I'm thinking one sesnor below the catapult to detect ball settled, and one on the top of the intake look ing downwards to detect that a ball has been brought all the way through the intake.