Modified ball sensors.
Changed winch subsystem to utilize two seperate sensors for detecting the presence of a ball.

One will be mounted on the intake looking athwartships, this will act as a beam breaker to determine when we have gotten a ball into the robot.
The other will be used to determine when a ball has settled on the catapult and is ready to be fired. It will be mounted under the catapult looking upwards.