Update ServoBallTapper.java
Command groups that used the servo tappers were never completing.
My guess is that this is because the commands attempting to move the tappers were never completing. Why? Because they were waiting for the tappers to actually reach the destination angle they were being commanded to. I assumed the servo class took in an angle normalized from 0 - 180 degrees, but if you look at the source, the range is actually defined as 0 - 170 degrees. Unfortunately they don't expose any methods by which you could change the range, so code needs to be modified to match their implementation. 

Still need to test if this resolves the problem. In the end it might be safer/faster to just return 'true' in isFinished for all the servo commands.

http://team2168.org/javadoc/src-html/edu/wpi/first/wpilibj/Servo.html#line.24